// src/app/game/enum/game-status.enum.ts
export enum GameStatus {
  PREPARE = 'prepare', // До запуска игры
  // слушаю событие запуска с кнопки и переключаю
  SITUATION = 'situation', // Вывод текста игры + Игроки бросают карты
  // жду последнего бросившего и переключаю
  VOTE = 'vote', // Голосование
  // жду последнего проголосовавшего и переключаю
  VOTE_RESULTS = 'vote-results', // Отображение результатов голосования
  // переключаю по таймеру
  END = 'end', // Финальный экран
  // жду, когда все покинут игру и удаляю лобби из мапы //TODO возможно потребуется отмена удаления при реконнекте, если игрок в лобби один (до начала игры)
}